pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mutual-fund-analyser"
version = "0.1.0"
description = "Web scraping and analysis pipeline for mutual fund holdings aggregation"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Rahul Garg" },
]
keywords = ["mutual-funds", "scraping", "playwright", "analysis", "finance"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Stock Investors",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

dependencies = [
  # scraping
  "playwright>=1.45.0,<2.0.0",
  "requests>=2.31.0,<3.0.0",

  # web dashboard
  "fastapi>=0.110.0,<1.0.0",
  "uvicorn>=0.29.0,<1.0.0",

  # data and utils
  "pydantic>=2.7.0,<3.0.0",
  "pydantic-settings>=2.2.0,<3.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "orjson>=3.10.3,<4.0.0",
  "pyyaml>=6.0.0,<7.0.0",
  "loguru>=0.7.2,<1.0.0",
  "pandas>=2.2.0,<3.0.0",
  
  # validation
  "jsonschema>=4.22.0,<5.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "pytest-mock>=3.14.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
  "types-PyYAML",
  "types-requests",
  "ipykernel",
  "pre-commit>=3.7.0",
]

[project.scripts]
mfa-analyze = "mfa.cli.analyze:main"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = [
  "--strict-markers",
  "--strict-config",
  "--cov=mfa",
  "--cov-report=term-missing",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__pycache__/*"]


[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

